requirements.txt
streamlit==1.29.0
pandas==2.1.4
gspread==5.12.4
google-auth==2.25.2
plotly==5.18.0
